% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barchart.R
\name{bar_fn}
\alias{bar_fn}
\title{Title}
\usage{
bar_fn(data)
}
\arguments{
\item{data}{this is a dataframe that you want to analyze and create barcharts for
to get more in depth analysis on the data}
}
\value{

}
\description{
Title
}
\examples{
bar_fn(mtcars)
}
