% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boxplot.R
\name{boxplot_fn}
\alias{boxplot_fn}
\title{Title}
\usage{
boxplot_fn(data)
}
\arguments{
\item{data}{this is a dataframe that you want to analyze and create boxplots for
to get more in depth analysis on the data}
}
\value{

}
\description{
Title
}
\examples{

boxplot_fn(mt,cars)
}
