% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hist.R
\name{hist_fn}
\alias{hist_fn}
\title{Title}
\usage{
hist_fn(data)
}
\arguments{
\item{data}{this is a dataframe that you want to analyze and create histograms for
to get more in depth analysis on the data}
}
\value{

}
\description{
Title
}
\examples{
hist_fn(mtcars)
}
