% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_imp.R
\name{med_imp}
\alias{med_imp}
\title{Title}
\usage{
med_imp(data)
}
\arguments{
\item{data}{this is a dataframe that you want to analyze and impute medians
where data may be missing}
}
\value{

}
\description{
Title
}
\examples{
med_imp(mtcars)
}
