% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drop_NAs.R
\name{NA_drop_rows}
\alias{NA_drop_rows}
\title{Title}
\usage{
NA_drop_rows(data, threshold)
}
\arguments{
\item{threshold}{a number between 0 and 1 that indicates at what level should
data be dropped if it constains too many NAs}

\item{data-}{this is a dataframe that you want to analyze to see if there are
any NAs}
}
\value{

}
\description{
Title
}
\examples{
NA_drop_rows(mtcars,0.15)
}
